input tagger for relevance tagging model (tags each word with the list of output statements it participates in)

+++ dynet/seqTagger/tagEnd.py
@@ -1,6 +1,10 @@
 import sys
 import os
-os.rename(sys.argv[1]+'.corpus.ll',sys.argv[1]+'.words')
+if not os.path.exists(sys.argv[1]+'.words'):
+    if os.path.exists(sys.argv[1]+'.ll'):
+        os.rename(sys.argv[1]+'.corpus.ll',sys.argv[1]+'.words')
+    else:
+        print 'Dataset input missing'
 with open(sys.argv[1]+'.words','r') as fin:
     with open(sys.argv[1]+'.tags','w') as fout:
         for l in fin.readlines():
